caption_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
transformers==4.44.0
pillow==10.4.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.4.0+cpu
